move module functions from operations.py to utils.py
